# inference: enable constant propagation for `invoke`d calls, fixes #41024