@nanosoldier 2021-06-27
[Your benchmark job](https://github.com/JuliaLang/julia/pull/41383) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/d6800b9_vs_05e4138/report.md). cc @christopher-dG