@aviatesk 2021-06-27
It turns out that this PR allows `@horner` to be constant-folded:
```julia
# master
julia> code_llvm(()) do
                  sin(sum(sincos(10)))
              end
;  @ REPL[16]:2 within `#23`
define double @"julia_#23_867"() #0 {
top:
  %0 = call double @j_sin_869(double 0xBFF62125BF7FDD38) #0
  ret double %0
}

# this PR
julia> code_llvm(()) do
           sin(sum(sincos(10)))
       end
;  @ REPL[12]:2 within `#19`
define double @"julia_#19_911"() #0 {
top:
  ret double 0xBFEF701C34D1CD3B
}
```

And so it wins some scalar benchmarks completely:
| ID | time ratio | memory ratio |
|----|------------|--------------|
| `["scalar", "sincos", ("argument reduction (easy) abs(x) < 2.0^20π/4", "negative argument", "Float64")]` | 0.00 (5%) :white_check_mark: | 1.00 (1%)  |
| `["scalar", "sincos", ("argument reduction (easy) abs(x) < 2.0^20π/4", "positive argument", "Float64")]` | 0.00 (5%) :white_check_mark: | 1.00 (1%)  |
| `["scalar", "sincos", ("argument reduction (easy) abs(x) < 2π/4", "negative argument", "Float32")]` | 0.00 (5%) :white_check_mark: | 1.00 (1%)  |
| `["scalar", "sincos", ("argument reduction (easy) abs(x) < 2π/4", "negative argument", "Float64")]` | 0.00 (5%) :white_check_mark: | 1.00 (1%)  |