@aviatesk 2021-06-27
Hmm, with certain compile time cost of course,...
```julia
# master
julia> @time using Plots
  3.564731 seconds (7.13 M allocations: 496.071 MiB, 6.93% gc time)

julia> @time plot(rand(10,3))
  2.662668 seconds (3.51 M allocations: 219.314 MiB, 5.52% gc time, 55.19% compilation time)

# this PR
ulia> @time using Plots
  3.810949 seconds (7.33 M allocations: 513.128 MiB, 6.46% gc time)

julia> @time plot(rand(10,3))
  2.867559 seconds (3.52 M allocations: 219.522 MiB, 5.04% gc time, 55.26% compilation time)
```