@aviatesk 2021-06-27
Okay 8028bc1 improves inference a bit and should fix the latency regression:
```julia
# master
julia> @time using Plots; @time plot(rand(10,3))
  4.126655 seconds (7.13 M allocations: 496.086 MiB, 7.48% gc time)
  2.908765 seconds (3.51 M allocations: 219.314 MiB, 4.60% gc time, 53.35% compilation time)

# after 8028bc1
julia> @time using Plots; @time plot(rand(10,3))
  3.889846 seconds (7.01 M allocations: 489.443 MiB, 7.06% gc time)
  2.918426 seconds (3.50 M allocations: 218.861 MiB, 4.42% gc time, 54.79% compilation time)
```