@aviatesk 2021-06-30
No, it only "fixes" the one specific case when we want LICM.
We should fix it more generally, i.e. even if a target component can't be constant-folded, we want to hoist it out of loop as far as it proven to be loop-invariant.