@JeffreySarnoff 2021-07-05
thank you, nicely done .

"We have the technology. We have the capability to build.. that.. Better than.. before. Better, stronger, faster."
    — The Six Million Dollar Man, Opening Narration [heavily redacted]